WikiParser/requirements.txt
charset-normalizer==2.0.9
mwparserfromhell==0.6.3
Pillow==9.0.1
psycopg2-binary==2.9.3
requests==2.27.1
